requirements.txt
streamlit==0.88.0
pandas==1.3.3
easyocr==1.4
mysql-connector-python==8.0.27
Pillow==8.3.2
opencv-python-headless==4.5.3.56
matplotlib==3.4.3
streamlit-option-menu==0.0.3